Moving map_caching.go from the otlp to the r8 package
